feat: Add render_to_formspec_string function
Abstracted:

- `Form:show` to seperate out the next_formspec logic into the new
  private `new_next_formname` function.
- `on_fs_input` to seperate out the callback logic into the new
  private `fs_process_events` function.

`Form:render_to_formspec_string` is a new public function, taking a
player ref and a context. Makes use of `new_next_formname` to get a
unique formname for debugging purposes. Returns a tuple of the
formspec and a asdfs function, which accepts a `fields` reference. This
function calls `fs_process_events` to trigger the form events, and
returns a boolean indicating if the form should be redrawn.